rearranging packages. sketching README
