+++ README.md
@@ -1,50 +1,32 @@
 # Zygo Programming Challenge - Android Developer
+Android challange implementado por José Diogo Castro
 
-Seu objetivo neste challenge é refatorar e adicionar novas features um pequeno aplicativo que lista Séries de TV.
+## Tecnologias adicionadas
+* Coroutines
+* Lifecycle
+* ViewModel
+* Koin
+* Chuck
+* Gson
+* ExpandableLayout (skydoves)
 
-## Especificações
+## Arquitetura
+* Mantive os 4 pacotes raiz `main`, `model`, `network` e `utils`. Cada um foi subdividos em outros pacotes como mostrado abaixo.
+* `main`: pacote principal da aplicação
+    * `adapters`: coletânea dos `RecyclerView.Adapter` e `RecyclerView.ViewHolder` da aplicação;
+    * `di.koin`: declaração dos módulos de Koin para DI;
+    * `ui.activity`: coletânea de `Activity` da aplicação;
+    * `viewModel`: coletânea de `ViewModel` da aplicação;
+* `model`: modelos da aplicação. Dividio em 4 subpacotes cada um referente a um tipo de entidade da aplicação: `episode`, `genre`, `season` e `show`.
+* `network`:  pacote responsável pela comunicação do app com a API e configurações do Retrofit. Dividido em 3 subpacotes:
+    * `api`: dividida em 2 subpacotes: `repository` e `service`;
+    * `model`: modelos da API e Retrofit (respostas e recursos);
+    * `retrofit`: configuração do cliente Retrofit para a API.
+* `utils`: pacote com classes genéricas, como *extensions functions* e Dialogs *factory*.
 
-O aplicativo possui apenas uma tela de listagem das séries, seu desafio consiste em:
+## Notas de design
 
-* Implementar a tela de Detalhes da Série;
-* Remover as chamadas de API da Main;
-* Implementar o novo layout seguindo as especificicações da nossa Designer;
-* Resolver o Crash que ocorre ao iniciar o aplicativo sem conexão à Internet;
-
-Acesse a [Documentação da API](https://developers.themoviedb.org/3/tv/get-tv-details) para qualquer dúvida!
-
-**Serão avaliados:** apenas a organização do código, uso
-das ferramentas disponíveis, conhecimento e domínio sobre as linguagens e a
-capacidade de implementação das especificações
-
-## Requisitos do Novo Layout
-As novas telas podem ser encontradas no [Invision](https://isabellataques225701.invisionapp.com/console/share/6Z2ABNOYVB/549307395)
-
-* A Série em destaque deve ser a top 1 mais popular retornada pela API.
-* O restante dos resultados devem ser organizados por Gênero como mostrado no Layout
-
-## Requisitos técnicos
-
-* O projeto atual foi desenvolvido em Kotlin, mas fique a vontade para converter para Java caso sinta-se mais confortável.
-* É permitido o uso de frameworks e bibliotecas externos, desde que dentro de um
-  sistema de gerenciamento de pacotes.
-* O código deve ser claro, preferencialmente documentado.
-* A arquitetura e design do sistema devem ser documentadas em um arquivo README
-  (brevemente, por favor).
-
-## Bônus!
-
-Testes são muito bem vindos, sobrando qualquer tempo, faça-os e ganhe uma
-pontuação extra!
-
-Se você tiver ainda mais tempo e quiser arriscar, temos uma listinha sobre o que pode ser melhorado no projeto:
-* Remover a necessidade de passar os parâmetros `api_key` e `region` para toda a chamada da API;
-* Armazenar a chave de API em um local seguro;
-* Implementar a busca por nome das Séries;
-
-## Envio
-
-Faça um fork desse repositório para o seu GitHub, crie um branch de desenvolvimento e faça todos os seus Commits nesse branch. Ao terminar abra um Pull Request para a branch main e nos envie o link do mesmo. 
+*
 
 ### Disclaimer
