+++ app/src/main/java/com/zygotecnologia/zygotv/main/adapters/DetailsAdapter.kt
@@ -7,7 +7,7 @@ import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.skydoves.expandablelayout.ExpandableLayout
 import com.zygotecnologia.zygotv.R
-import com.zygotecnologia.zygotv.model.SeasonResponse
+import com.zygotecnologia.zygotv.model.season.SeasonResponse
 import com.zygotecnologia.zygotv.utils.ImageUrlBuilder.loadImage
 import com.zygotecnologia.zygotv.utils.gone
 import com.zygotecnologia.zygotv.utils.inflate
