+++ app/src/main/java/com/zygotecnologia/zygotv/main/di/koin/MainModule.kt
@@ -5,17 +5,15 @@ import com.zygotecnologia.zygotv.BuildConfig
 import com.zygotecnologia.zygotv.main.viewModel.DetailsViewModel
 import com.zygotecnologia.zygotv.main.viewModel.MainViewModel
 import com.zygotecnologia.zygotv.network.api.repository.ApiRepository
-import com.zygotecnologia.zygotv.network.api.repository.TmdbApiRepository
+import com.zygotecnologia.zygotv.network.api.repository.ApiRepositoryImpl
 import com.zygotecnologia.zygotv.network.api.service.ApiService
 import com.zygotecnologia.zygotv.network.retrofit.OkHttpClientFactory
 import com.zygotecnologia.zygotv.network.retrofit.RetrofitClient
-import com.zygotecnologia.zygotv.utils.DialogFactory
 import okhttp3.Interceptor
 import okhttp3.logging.HttpLoggingInterceptor
 import org.koin.android.ext.koin.androidContext
 import org.koin.androidx.viewmodel.dsl.viewModel
 import retrofit2.Retrofit
-import retrofit2.converter.moshi.MoshiConverterFactory
 
 
 const val fullBaseUrl = BuildConfig.BASE_URL + "/" + BuildConfig.API_VERSION + "/"
@@ -30,7 +28,7 @@ val MainModule = module {
 
     // API
     single { get<Retrofit>().create(ApiService::class.java) }
-    factory<ApiRepository> { TmdbApiRepository(get(), get()) }
+    factory<ApiRepository> { ApiRepositoryImpl(get(), get()) }
 
     // ViewModels
     viewModel { MainViewModel(get()) }
