+++ app/src/main/java/com/zygotecnologia/zygotv/main/viewModel/MainViewModel.kt
@@ -3,11 +3,11 @@ package com.zygotecnologia.zygotv.main.viewModel
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.zygotecnologia.zygotv.model.Genre
-import com.zygotecnologia.zygotv.model.Show
+import com.zygotecnologia.zygotv.model.genre.Genre
+import com.zygotecnologia.zygotv.model.show.Show
 import com.zygotecnologia.zygotv.network.api.repository.ApiRepository
-import com.zygotecnologia.zygotv.network.retrofit.model.RequestError
-import com.zygotecnologia.zygotv.network.retrofit.model.RetrofitResponse
+import com.zygotecnologia.zygotv.network.model.RequestError
+import com.zygotecnologia.zygotv.network.model.RetrofitResponse
 import kotlinx.coroutines.launch
 
 class MainViewModel(private val api: ApiRepository) : ViewModel() {
