+++ app/src/main/java/com/zygotecnologia/zygotv/model/genre/GenreResponse.kt
@@ -1,0 +1,11 @@
+package com.zygotecnologia.zygotv.model.genre
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class GenreResponse(
+    @SerializedName("genres")
+    val genres: List<Genre>?
+) : Parcelable
