+++ app/src/main/java/com/zygotecnologia/zygotv/model/season/SeasonResponse.kt
@@ -1,19 +1,20 @@
-package com.zygotecnologia.zygotv.model
+package com.zygotecnologia.zygotv.model.season
 
 import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
+import com.zygotecnologia.zygotv.model.episode.Episode
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
-data class Season(
-    @SerializedName("id")
-    val id: Int?,
-
+data class SeasonResponse(
     @SerializedName("season_number")
     val seasonNumber: Int?,
 
     @SerializedName("name")
     val name: String?,
+
+    @SerializedName("episodes")
+    val episodes: List<Episode>?,
 
     @SerializedName("overview")
     val overview: String?,
