+++ app/src/main/java/com/zygotecnologia/zygotv/network/model/RequestError.kt
@@ -1,4 +1,4 @@
-package com.zygotecnologia.zygotv.network.retrofit.model
+package com.zygotecnologia.zygotv.network.model
 
 import android.os.Parcelable
 import com.squareup.moshi.Json
