+++ app/src/main/java/com/zygotecnologia/zygotv/network/retrofit/RetrofitExtensions.kt
@@ -1,8 +1,8 @@
 package com.zygotecnologia.zygotv.network.retrofit
 
 import androidx.lifecycle.MutableLiveData
-import com.zygotecnologia.zygotv.network.retrofit.model.RequestError
-import com.zygotecnologia.zygotv.network.retrofit.model.Resource
+import com.zygotecnologia.zygotv.network.model.RequestError
+import com.zygotecnologia.zygotv.network.model.Resource
 
 fun <T> Resource<T>.validateResponse(success: MutableLiveData<T>, error: MutableLiveData<RequestError>) {
     when (this.status) {
